+++ src/caching/GlobalCache.ts
@@ -9,7 +9,7 @@ export class $ { // $ = cash = cache... get it? :D
 
 	static structures<T extends Structure>(saver: { ref: string }, key: string, callback: () => T[],
 										   timeout = CACHE_TIMEOUT): T[] {
-		let cacheKey = saver.ref + ':' + key;
+		const cacheKey = saver.ref + ':' + key;
 		if (!_cache.structures[cacheKey] || Game.time > _cache.expiration[cacheKey]) {
 			// Recache if new entry or entry is expired
 			_cache.structures[cacheKey] = callback();
@@ -26,7 +26,7 @@ export class $ { // $ = cash = cache... get it? :D
 	}
 
 	static number(saver: { ref: string }, key: string, callback: () => number, timeout = SHORT_CACHE_TIMEOUT): number {
-		let cacheKey = saver.ref + ':' + key;
+		const cacheKey = saver.ref + ':' + key;
 		if (_cache.numbers[cacheKey] == undefined || Game.time > _cache.expiration[cacheKey]) {
 			// Recache if new entry or entry is expired
 			_cache.numbers[cacheKey] = callback();
@@ -36,7 +36,7 @@ export class $ { // $ = cash = cache... get it? :D
 	}
 
 	static list<T>(saver: { ref: string }, key: string, callback: () => T[], timeout = CACHE_TIMEOUT): T[] {
-		let cacheKey = saver.ref + ':' + key;
+		const cacheKey = saver.ref + ':' + key;
 		if (_cache.lists[cacheKey] == undefined || Game.time > _cache.expiration[cacheKey]) {
 			// Recache if new entry or entry is expired
 			_cache.lists[cacheKey] = callback();
@@ -47,7 +47,7 @@ export class $ { // $ = cash = cache... get it? :D
 
 	static costMatrix(roomName: string, key: string, callback: () => CostMatrix,
 					  timeout = SHORT_CACHE_TIMEOUT): CostMatrix {
-		let cacheKey = roomName + ':' + key;
+		const cacheKey = roomName + ':' + key;
 		if (_cache.costMatrices[cacheKey] == undefined || Game.time > _cache.expiration[cacheKey]) {
 			// Recache if new entry or entry is expired
 			_cache.costMatrices[cacheKey] = callback();
@@ -64,7 +64,7 @@ export class $ { // $ = cash = cache... get it? :D
 	static set<T extends HasRef, K extends keyof T>(thing: T, key: K,
 													callback: () => (T[K] & (undefined | HasID | HasID[])),
 													timeout = CACHE_TIMEOUT) {
-		let cacheKey = thing.ref + '$' + key;
+		const cacheKey = thing.ref + '$' + key;
 		if (!_cache.things[cacheKey] || Game.time > _cache.expiration[cacheKey]) {
 			// Recache if new entry or entry is expired
 			_cache.things[cacheKey] = callback();
@@ -72,11 +72,11 @@ export class $ { // $ = cash = cache... get it? :D
 		} else {
 			// Refresh structure list by ID if not already done on current tick
 			if (_cache.accessed[cacheKey] < Game.time) {
-				if (_.isArray(_cache.things[cacheKey])) {
+				if (!_.isArray(_cache.things[cacheKey])) {
+					_cache.things[cacheKey] = Game.getObjectById((<HasID>_cache.things[cacheKey]).id) as HasID;
+				} else {
 					_cache.things[cacheKey] = _.compact(_.map(_cache.things[cacheKey] as HasID[],
 															  s => Game.getObjectById(s.id))) as HasID[];
-				} else {
-					_cache.things[cacheKey] = Game.getObjectById((<HasID>_cache.things[cacheKey]).id) as HasID;
 				}
 				_cache.accessed[cacheKey] = Game.time;
 			}