+++ src/prototypes/Creep.ts
@@ -25,3 +25,9 @@ Object.defineProperty(Creep.prototype, 'boostCounts', {
 	configurable: true,
 });
 
+Object.defineProperty(Creep.prototype, 'inRampart', {
+	get() {
+		return !!this.pos.lookForStructure(STRUCTURE_RAMPART); // this assumes hostile creeps can't stand in my ramparts
+	},
+	configurable: true,
+});